requirements.txt
langchain>=0.2.0
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
sentence-transformers
pgvector
psycopg2-binary
pypdf
python-dotenv

fastapi
uvicorn
passlib[bcrypt]
pydantic[email]

mcp-use